refactor: restructure codebase directories and filename conventions for enhanced organization
